Updates RainbowKit configuration as described at https://www.rainbowkit.com/docs/migration-guide and structure of project to next 13 as described at https://nextjs.org/docs/app/building-your-application/upgrading/app-router-migration#step-3-migrating-nexthead
