+++ packages/react-app/app/page.tsx
@@ -1,38 +1,10 @@
-'use client';
+import HomePage from './home-page'
+import AppProvider from '../providers/AppProvider'
 
-import { useEffect, useState } from 'react';
-
-import { useAccount } from 'wagmi';
-
-export default function Home() {
-  const [userAddress, setUserAddress] = useState('');
-  const [isMounted, setIsMounted] = useState(false);
-  const { address, isConnected } = useAccount();
-
-  useEffect(() => {
-    setIsMounted(true);
-  }, []);
-
-  useEffect(() => {
-    if (isConnected && address) {
-      setUserAddress(address);
-    }
-  }, [address, isConnected]);
-
-  if (!isMounted) {
-    return null;
-  }
-
+export default function Page () {
   return (
-    <div className="flex flex-col justify-center items-center">
-      <div className="h1">
-        There you go... a canvas for your next Celo project!
-      </div>
-      {isConnected ? (
-        <div className="h2 text-center">Your address: {userAddress}</div>
-      ) : (
-        <div>No Wallet Connected</div>
-      )}
-    </div>
+    <AppProvider>
+      <HomePage />
+    </AppProvider>
   );
 }
