+++ packages/react-app/providers/AppProvider.tsx
@@ -1,48 +1,35 @@
 'use client';
 
 import '@rainbow-me/rainbowkit/styles.css';
-
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import {
+  getDefaultConfig,
   RainbowKitProvider,
-  connectorsForWallets,
 } from '@rainbow-me/rainbowkit';
-import { WagmiProvider, createConfig, http } from 'wagmi';
-import { celo, celoAlfajores } from 'wagmi/chains';
+import { WagmiProvider } from 'wagmi';
+import {
+  celo,
+  celoAlfajores
+} from 'wagmi/chains';
 
-import Layout from '../components/Layout';
-import { injectedWallet } from '@rainbow-me/rainbowkit/wallets';
-
-const connectors = connectorsForWallets(
-  [
-    {
-      groupName: 'Recommended',
-      wallets: [injectedWallet],
-    },
+const config = getDefaultConfig({
+  appName: 'RainbowKit App',
+  projectId: 'YOUR_PROJECT_ID',
+  chains: [
+    celo,
+    ...(process.env.NEXT_PUBLIC_ENABLE_TESTNETS === 'true' ? [celoAlfajores] : []),
   ],
-  {
-    appName: 'Celo Composer',
-    projectId: process.env.WC_PROJECT_ID ?? '044601f65212332475a09bc14ceb3c34',
-  }
-);
-
-const config = createConfig({
-  connectors,
-  chains: [celo, celoAlfajores],
-  transports: {
-    [celo.id]: http(),
-    [celoAlfajores.id]: http(),
-  },
+  ssr: true,
 });
 
 const queryClient = new QueryClient();
 
-export function AppProvider({ children }: { children: React.ReactNode }) {
+export default function AppProvider({ children }: { children: React.ReactNode }) {
   return (
     <WagmiProvider config={config}>
       <QueryClientProvider client={queryClient}>
         <RainbowKitProvider>
-          <Layout>{children}</Layout>
+          {children}
         </RainbowKitProvider>
       </QueryClientProvider>
     </WagmiProvider>
